superheroes/superhero-02-02/requirements.txt
requests==2.32.3
python-dotenv==1.0.1
coloredlogs==15.0.1
jwt==1.3.1
flask==2.2.5
flask-restful==0.3.10
google-generativeai==0.8.3